правки замечаний
