+++ src/components/button/button.module.css
@@ -1,0 +1,9 @@
+.button {
+  cursor: pointer;
+  background-color: var(--body-bkg-color);
+  color: var(--text-color);
+  border-radius: 4px;
+  border: 1px solid gray;
+  padding: 3px 5px;
+  margin: 0 5px;
+}
