+++ src/components/login-context/hooks.js
@@ -1,4 +1,4 @@
 import { useContext } from "react";
 import { LoginContext } from "./login-context-provider.jsx";
 
-export const useLogin = () => useContext(LoginContext);
+export const useLoginContext = () => useContext(LoginContext);
