+++ src/components/login-context/login-context-provider.jsx
@@ -5,5 +5,11 @@ export const LoginContext = createContext();
 export const LoginContextProvider = ({ children }) => {
   const [user, setUser] = useState();
 
-  return <LoginContext value={{ user, setUser }}>{children}</LoginContext>;
+  const login = (name = "Пупкин") => setUser(name);
+
+  const logout = () => setUser("");
+
+  return (
+    <LoginContext value={{ user, login, logout }}>{children}</LoginContext>
+  );
 };
