+++ src/components/restaurant-tile/restaurant-tile.jsx
@@ -2,10 +2,10 @@ import RestaurantComments from "../restaurant-comments/restaurant-comments.jsx";
 import RestaurantMenu from "../restaurant-menu/restaurant-menu.jsx";
 import { ReviewForm } from "../review-form/review-form.jsx";
 import styles from "./restaurant-tile.module.css";
-import { useLogin } from "../login-context/login-context-hooks.js";
+import { useLoginContext } from "../login-context/hooks.js";
 
 function RestaurantTile({ name, menu, reviews }) {
-  const { user } = useLogin();
+  const { user } = useLoginContext();
   return (
     <div className={styles.container}>
       <h2 className={styles.name}>{name}</h2>
