--- src/components/theme-context/theme-context-hooks.js
@@ -1,4 +1,0 @@
-import { useContext } from "react";
-import { ThemeContext } from "./theme-context-provider.jsx";
-
-export const useTheme = () => useContext(ThemeContext);
