+++ src/main.css
@@ -42,13 +42,3 @@ h3 {
     line-height: 1.5em;
     padding-bottom: 0.5em;
 }
-
-.button {
-    cursor: pointer;
-    background-color: var(--body-bkg-color);
-    color: var(--text-color);
-    border-radius: 4px;
-    border: 1px solid gray;
-    padding: 3px 5px;
-    margin: 0 5px;
-}
